+++ test/Confluent.Kafka.IntegrationTests/testconf.json
@@ -1,3 +1,4 @@
 {
-  "bootstrapServers": "127.0.0.1:9092"
+  "bootstrapServers": "127.0.0.1:9092",
+  "oauthbearerBootstrapServers": "127.0.0.1:9093"
 }
